chore: move `receiveClientRequest()` and `dispatchRequest()` to separate files
