+++ src/sockets-polling/receive-client-request.cpp
@@ -12,76 +12,7 @@ int handleParsingError(const HttpException& e, HttpResponse& response, Client& c
     return REQUEST_ERR;
 }
 
-void dispatchRequest(Client& client)
-{
-    HttpRequest& request = client.request;
-    HttpResponse& response = client.response;
-
-    response.setHeader("Connection", request.getHeader("Connection"));
-    if (request.getHeader("Connection") == "close")
-        client.connType = CLOSE;
-
-    if (!client.location.getCgiPath().empty()) {
-        std::cout << GREEN "CGI found\n" RESET;
-        Cgi cgi;
-
-        std::cout << "response body: \n" << response.getBody() << '\n';
-        std::string cgiOutput = cgi.executeCgi(request, response);
-        std::cerr << "cgiOutput: " << cgiOutput << "\n";
-
-        if (cgiOutput.empty())
-            throw HttpException(INTERNAL_ERROR, "CGI script execution failed");
-        else {
-            handleCgiRequest(cgiOutput, response);
-            std::cout << "---------- CGI Output ----------\n" << BLUE << response.toString() << RESET << "\n";
-        }
-    }
-    else {
-        if (request.getMethod() == "GET")
-        {
-            std::cout << GREEN "GET request\n" RESET;
-            response.handleGetRequest(client.location, client);
-        }
-        else if (request.getMethod() == "POST")
-        {
-            std::cout << GREEN "POST request\n" RESET;
-            response.handlePostRequest(request, client);
-        }
-        else if (request.getMethod() == "DELETE")
-            throw HttpException(METHOD_NOT_ALLOWED, "Delete without CGI not allowed");
-    }
-}
-
-void acceptClient(std::vector<struct pollfd>& pfds, std::vector<Client>& clients, int listener)
-{
-    std::cout << "POLLIN: socket " << listener << '\n';
-
-    // If listener is ready to read, handle new client
-    struct sockaddr_storage remoteAddr; // Client address
-    socklen_t addrLen = sizeof remoteAddr;
-    int newFd = accept(listener, (struct sockaddr *)&remoteAddr, &addrLen);
-    if (newFd == -1) {
-        perror("accept");
-        return ;
-    }
-    if (set_nonblocking(newFd) == -1) {
-        perror("set_nonblocking (new_fd)");
-        close(newFd);
-        return ;
-    }
-
-    addToPfds(pfds, newFd);
-    clients.push_back(Client(newFd, getNowInSeconds()));
-
-    char remoteIp[INET6_ADDRSTRLEN];
-    printf("server: new client from %s on "
-        "socket %d\n", inet_ntop(remoteAddr.ss_family,
-            getInAddr((struct sockaddr*)&remoteAddr),
-            remoteIp, INET6_ADDRSTRLEN),
-        newFd);
-}
-
-void validateMethod(const std::string& method, const std::vector<std::string>& allowedMethods)
+static void validateMethod(const std::string& method, const std::vector<std::string>& allowedMethods)
 {
     for (std::vector<std::string>::const_iterator It = allowedMethods.begin(); It != allowedMethods.end(); ++It) {
         if (*It == method)
@@ -89,8 +20,6 @@ void validateMethod(const std::string& method, const std::vector<std::string>& a
     }
     throw HttpException(METHOD_NOT_ALLOWED, "Method not allowed");
 }
-
-
 
 
 /*
