+++ src/sockets-polling/timeout.cpp
@@ -1,5 +1,10 @@
 #include "../../include/sockets-polling.h"
 
+time_t getNowInSeconds() {
+    struct timeval tv;
+    gettimeofday(&tv, NULL);
+    return tv.tv_sec;
+}
 
 static time_t getTimeoutBySocketFd(int fd, std::map< std::pair<std::string, std::string> , std::vector<Server> >& servers)
 {
